// packages/ui/src/common/enums/auto-start-chat.ts
export enum AUTO_CHAT_ERROR_EXIT_CODE {
  NORMAL = 0,
  COOKIE_INVALID = 81,
  LOGIN_STATUS_INVALID = 82,
  ERR_INTERNET_DISCONNECTED = 83,
  ACCESS_IS_DENIED = 84,
  PUPPETEER_IS_NOT_EXECUTABLE = 85,
  LLM_UNAVAILABLE = 86,
}

export enum RECHAT_CONTENT_SOURCE {
  LOOK_FORWARD_EMOTION = 1,
  GEMINI_WITH_CHAT_CONTEXT = 2
}

export enum RECHAT_LLM_FALLBACK {
  SEND_LOOK_FORWARD_EMOTION = 1,
  EXIT_REMINDER_PROGRAM = 2
}

export enum RUNNING_STATUS_ENUM {
  RUNNING = 0,
  NORMAL_EXITED = 1,
  ERROR_EXITED = 2
}